add esp32 code, rename class for esp8266
